fixed: refactoring
