+++ frontend/src/components/query-input/create-speech-recognition.js
@@ -1,0 +1,3 @@
+export function createSpeechRecognition() {
+    return window.speechRecognition || window.webkitSpeechRecognition
+}
