+++ frontend/src/index.html
@@ -2,6 +2,7 @@
 <html lang="en">
 <head>
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <link rel="shortcut icon" href="./public/img/favicon.ico">
   <script src="https://code.responsivevoice.org/responsivevoice.js?key=oZmnbAtI"></script>
   <script type="importmap">
     {
@@ -30,7 +31,7 @@
 <!--  </div>-->
 
   <div id="avatarContainer" style="background-color:white; display: flex; justify-content: center; align-items: center">
-    <img src="src/video/seamless-loop-talking-unscreen.gif" alt="EISBÄR" width="60%" id="2dAvatar">
+    <img src="./public/video/seamless-loop-talking-unscreen.gif" alt="EISBÄR" width="60%" id="2dAvatar">
   </div>
